+++ lib/constants/routes.dart
@@ -1,6 +1,6 @@
 import 'package:to_do_list_project/views/tasks/list_tasks_page.dart';
 import 'package:to_do_list_project/views/my_home.dart';
-import 'package:to_do_list_project/views/tasks/new_task_page.dart';
+import 'package:to_do_list_project/views/tasks/task_page.dart';
 
 final myRoutes ={ 
   'home' : (context) => const MyHomePage(),
