+++ lib/views/my_home.dart
@@ -1,76 +1,113 @@
 import 'package:flutter/material.dart';
 
-
 class MyHomePage extends StatelessWidget {
   const MyHomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Home'),
-      ),
-      body: Center(
-        child: SizedBox(
-          width: MediaQuery.of(context).size.width * .8,
-          child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Container(
-                  height: MediaQuery.of(context).size.height * .20,
-                  decoration: const BoxDecoration(
-                      image: DecorationImage(
-                    image: AssetImage("assets/images/taskico.png"),
-                  )),
-                ),
-                SizedBox(
-                  height: 120,
-                  child: ElevatedButton(
-                    onPressed: (() {
-                      Navigator.pushNamed(context, 'new_task');
-                    }),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: const [
+      body: Container(
+        color: Colors.indigo,
+        child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+          Container(
+            height: MediaQuery.of(context).size.height * .30,
+            decoration: const BoxDecoration(
+                image: DecorationImage(
+              fit: BoxFit.fitHeight,
+              filterQuality: FilterQuality.high,
+              isAntiAlias: true,
+              image: AssetImage("assets/images/taskhome.png"),
+            )),
+          ),
+          const SizedBox(
+            height: 20,
+          ),
+          const Text(
+            'To Do List App',
+            style: TextStyle(fontSize: 35, color: Colors.white, fontStyle: FontStyle.italic),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(
+            height: 20,
+          ),
+          SizedBox(
+            height: 100,
+            width: MediaQuery.of(context).size.width * .8,
+            child: ElevatedButton(
+              onPressed: (() {
+                Navigator.pushNamed(context, 'new_task');
+              }),
+              child: Row(children: [
+                Expanded(
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: const [
                       Text(
                         'New Task',
-                        style: TextStyle(fontSize: 20),
+                        style: TextStyle(fontSize: 25),
                       ),
-                      SizedBox(width: 10,),
-                      Icon(Icons.add, size: 35,),
-                    ]),
+                      Text(
+                        'You can organize your tasks \nby adding your tasks into separate categories',
+                        style:
+                            TextStyle(fontSize: 13, fontStyle: FontStyle.italic),
+                      ),
+                    ],
                   ),
                 ),
                 const SizedBox(
-                  height: 10,
+                  width: 10,
                 ),
-                SizedBox(
-                  height: 120,
-                  child: ElevatedButton(
-                    onPressed: (() {
-                      Navigator.pushNamed(context, 'list_task');
-                    }),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: const[
-                       Text(
-                          'List Task',
-                          style: TextStyle(fontSize: 20),
-                        ),
-                        SizedBox(width: 10,),
-                        Icon(Icons.edit_note_outlined, size: 35,),
-                      ],
-                    ),
-                    
-                  ),
-                  
-                ),
-                const SizedBox(
-                  height: 10,
+                const Icon(
+                  Icons.add,
+                  size: 35,
                 ),
               ]),
-        ),
+            ),
+          ),
+          const SizedBox(
+            height: 10,
+          ),
+          SizedBox(
+            height: 100,
+            width: MediaQuery.of(context).size.width * .8,
+            child: ElevatedButton(
+              onPressed: (() {
+                Navigator.pushNamed(context, 'list_task');
+              }),
+              child: Row(
+                children: [
+                  Expanded(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: const [
+                        Text(
+                          'List Task',
+                          style: TextStyle(fontSize: 25),
+                        ),
+                        Text(
+                          'You can check all your pending/running/in progress tasks here',
+                          style: TextStyle(fontSize: 13, fontStyle: FontStyle.italic),
+                        ),
+                      ],
+                    ),
+                  ),
+                  const SizedBox(
+                    width: 10,
+                  ),
+                  const Icon(
+                    Icons.edit_note_outlined,
+                    size: 35,
+                  ),
+                ],
+              ),
+            ),
+          ),
+          const SizedBox(
+            height: 10,
+          ),
+        ]),
       ),
     );
   }
