+++ lib/views/tasks/list_tasks_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:to_do_list_project/constants/status_task.dart';
 import 'package:to_do_list_project/repository/task_repository.dart';
-import 'package:to_do_list_project/views/tasks/new_task_page.dart';
+import 'package:to_do_list_project/views/tasks/task_page.dart';
 
 import '../../model/task_model.dart';
 
@@ -14,10 +14,15 @@ class ListTaskPage extends StatefulWidget {
 
 class _ListTaskPageState extends State<ListTaskPage> {
   List<TaskModel> listTasks = [];
+  bool isLoading = true;
 
   void getTasks() async {
+    isLoading = true;
+    listTasks=[];
     listTasks = await TaskRepository().getAllTasks() ?? [];
-    setState(() {});
+    setState(() {
+      isLoading=false;
+    });
   }
 
   @override
@@ -29,8 +34,8 @@ class _ListTaskPageState extends State<ListTaskPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Your tasks')),
-      body: ListView.builder(
+      appBar: AppBar(title: const Text('Your personal tasks')),
+      body:isLoading==true? const Center(child: CircularProgressIndicator(),) : ListView.builder(
           itemCount: listTasks.length,
           itemBuilder: ((context, index) {
             String statusIndexTask = '';
@@ -52,34 +57,30 @@ class _ListTaskPageState extends State<ListTaskPage> {
                     color: Colors.orange, size: 50);
                 break;
             }
-            return SizedBox(
-              width: MediaQuery.of(context).size.width * .75,
-              height: 100,
-              child: Card(
-                shape: const RoundedRectangleBorder(),
-                elevation: 8,
-                child: ListTile(
-                  onTap: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) => TaskPage(
-                                task: listTasks[index],
-                              )),
-                    );
-                  },
-                  leading: iconIndexedTask,
-                  title: Text(listTasks[index].titleTask.toUpperCase(),
-                      style: Theme.of(context).textTheme.titleLarge),
-                  subtitle: Text(
-                      '${listTasks[index].descriptionTask.toString()} \n$statusIndexTask',
-                      style: Theme.of(context).textTheme.titleMedium),
-                  trailing: const Icon(Icons.edit_note_rounded),
-                  isThreeLine: true,
-                ),
-              ),
+            return Card(
+              shape: const RoundedRectangleBorder(),
+              elevation: 8,
+              child: ListTile(
+                onTap: () async {
+                  await Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                        builder: (context) => TaskPage(
+                              task: listTasks[index],
+                            )),
+                  );
+                  getTasks();
 
-              //         Text(statusIndexTask, style: Theme.of(context).textTheme.titleMedium),
+                },
+                leading: iconIndexedTask,
+                title: Text(listTasks[index].titleTask.toUpperCase(),
+                    style: Theme.of(context).textTheme.titleLarge),
+                subtitle: Text(
+                    '${listTasks[index].descriptionTask.toString()} \nSTATUS: $statusIndexTask',
+                    style: Theme.of(context).textTheme.titleMedium),
+                trailing: const Icon(Icons.edit_note_rounded),
+                isThreeLine: true,
+              ),
             );
           })),
     );
